% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacProdxy.R
\docType{data}
\name{bacProdxy}
\alias{bacProdxy}
\title{Bacterial production data set}
\format{A data frame with the coordinates of the 25 sampled locations}
\usage{
data(bacProdxy)
}
\description{
Longitude and latitude of 25 samples when bacterial production was sampled in
Lake St. Pierre (Québec, Canada). These sampled were carried out in August
18, 2005.
}
\references{
Blanchet F.G., P. Legendre, R. Maranger, D. Monti, and P. Pepin.
  (2011) Modelling the effect of directional spatial ecological processes at
  different scales. \emph{Oecologia}, 166, 357-368.
}
\keyword{datasets}

