% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlistw.R
\name{createlistw}
\alias{createlistw}
\title{Interactive tool to generate R code that creates a spatial weighting matrix}
\usage{
createlistw()
}
\description{
Interactive tool to generate R code that creates a spatial weighting matrix
}
\examples{
\dontrun{
## a matrix or an object of class 'Spatial*' should be in the global environment
xy <- matrix(rnorm(50), 25)
createlistw()
}
}
\seealso{
\code{\link{chooseCN}}
}
\author{
Stéphane Dray \email{stephane.dray@univ-lyon1.fr}
}
