% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listw.select.R
\name{listw.select}
\alias{listw.select}
\alias{mem.select}
\alias{mem.select}
\title{Function to optimize the selection of a spatial weighting matrix and select the best 
subset of eigenvectors}
\usage{
listw.select(x, candidates, autocor = c("positive", "negative", "all"),
  crit = c("forward", "MIR", "global"), alpha = 0.05, correction = TRUE,
  summary = TRUE)

mem.select(x, candidates, autocor = c("positive", "negative", "all"),
  crit = c("forward", "MIR", "global"), alpha = 0.05, summary = TRUE)
}
\arguments{
\item{x}{Vector, matrix, or dataframe of the response variable(s)}

\item{candidates}{A list of one or more spatial weighting matrices of the class 
\code{listw}; \code{candidates} can be a list containing one or several spatial 
weighting matrices, but it can also be a single \code{listw} object}

\item{autocor}{Sign of the spatial eigenvectors to consider; "positive", "negative", 
or "all", for positively, negatively autocorrelated eigenvectors, or both, respectively; 
default is "positive"}

\item{crit}{Criterion to select the best SWM among the set of significant candidate 
matrices. Either \code{forward} (default option), \code{"MIR"} (for univariate \code{x} 
only), or \code{"global"} (see \code{Details})}

\item{alpha}{Uncorrected predefined significance value; default is 0.05}

\item{correction}{wheter the p-value of the global test performed on each SWM should
be corrected for multiple tests (TRUE) or not (FALSE); default is \code{TRUE}}

\item{summary}{Logical; Whether a message summarizing the results should be returned
or not; The R2 value returned in the message depends on the argument \code{crit} (R2 of 
the subset of selected predictors or R2 of the global model); Default is \code{TRUE}}
}
\value{
\code{listw.select} returns two lists containing several features of the SWM selected
by the optimization procedure, and general features of all the SWMs generated and tested: 

The first list, \code{$all}, contains: 

\describe{
\item{MEM.all}{A list of all the complete set of MEM variables (as generated by 
\code{scores.listw}) for the candidate SWMs provided to the function, in the same order as 
the SWMs in \code{candidates}.}
\item{pval}{The p-values (corrected by the Sidak correction if \code{correction = TRUE})
of the global models for each SWM. If \code{autocor = "all"}, two values per SWM (positive 
and negative MEM models).}
\item{R2.global}{The adjusted R2 of the complete set of MEM variables of each SWM.
R2.global is provided only for the significant SWMs. If \code{autocor = "all"}, two values 
per SWM (positive If \code{autocor = "all"}, two values per SWM (positive 
and negative MEM models).and negative MEM models).}
}

The second list, \code{$best}, contains the features of the best model: 

\describe{
\item{MEM.all}{Dataframe of the complete set of eigenvectors of the best SWM.}
\item{MEM.select}{Dataframe of the subset of eigenvectors selected by the criterion
\code{crit}. This set of spatial predictors is the one to be used in further analyses 
(e.g. variation partitioning, ordinary least squares, GLM) (see \code{Details} section). 
If \code{crit = "global"}, then \code{MEM.select} contains the entire set of MEM variables 
generated for the corresponding SWM. In this case, \code{MEM.select} and \code{MEM.all}
are the same.} 
\item{listw}{Element of \code{candidates} corresponding to the best SWM. It is an 
object of class \code{listw}.}
\item{MEM.AdjR2Cum}{Only for \code{crit = "forward"}; Vector of the cumulative adjusted 
R2 of the eigenvectors selected by the forward selection (\code{MEM.select}).}
\item{name}{Name of the best SWM ("Connectivity matrix_Weighting matrix") if the 
list of candidate SWMs was generated by the function \code{\link{listw.candidates}}. 
Otherwise, corresponds to the index of the best SWM in the list \code{candidates} 
provided to \code{listw.select}.} 
\item{pval}{Global p-value of the best SWM (after multiple-test correction, if 
the \code{correction} argument = TRUE). If \code{autocor = "all"}, two values (positive and 
negative MEM models).}
\item{R2.global}{The adjusted R2 of the complete set of MEM variables of the best W 
matrix. If \code{autocor = "all"}, two values (positive and negative MEM models).}
\item{R2.select}{Adjusted R2 of \code{MEM.select}. If \code{crit = "global"}, then 
\code{R2.select} is equal to \code{R2.global}. If \code{autocor = "all"}, two values 
(positive and negative MEM models).}
\item{NbVar}{Number of spatial predictors selected by the forward selection. If 
\code{autocor = "all"}, two values (positive and negative MEM models).}
\item{bestw_index}{Index of the best SWM in the list \code{candidates} provided 
to \code{listw.select}.}
}

The function \code{mem.select} only returns one list, corresponding to the \code{$best} list
of \code{listw.selec}, but without the objects \code{name} and \code{bestW_index}.
}
\description{
\code{listw.select} computes MEM variables (i.e., eigenvectors of a doubly centered spatial 
weighting matrix) for various definitions of spatial weighting matrices (SWM) and optimizes 
the selection of the SWM and of a subset of MEM variables. 
The optimization is done by maximizing the adjusted R-squared (R2) or by minimizing the 
residual spatial autocorrelation. The function controls the type I error rate by accounting 
for the number of tests performed. 
Function \code{mem.select} computes MEM variables for a single definition of SWM, test its 
significances, and optimizes the selection of MEM variables within the SWM using the same
criteria than \code{listw.select}.
These functions combine calls to the functions \code{scores.listw} and \code{forward.sel} 
or \code{\link{MEM.moransel}}.
The list of candidate SWMs can easily be generated using \code{\link{listw.candidates}}.
The significance of each SWM is tested by 9999 permutations with the function 
\code{anova.cca} (package \code{vegan}).
}
\details{
While the selection of the SWM is the most critical step of the spatial
eigenvector-based methods (Dray et al. 2006), Bauman et al. (2018a) showed that 
optimizing the choice of the SWM led to inflated type I error rates if an
explicit control of the number of SWMs tested was not applied. The function
\code{listw.select} therefore applies a Sidak correction (Sidak 1967) for multiple tests to
the p-value of the global test of each SWM (i.e., the model integrating the 
whole set of spatial predictors). The Sidak correction is computed as: 
\eqn{P_corrected = 1 - (1 - P)^n}, where \eqn{n} is the number of tests performed, \eqn{P}
is the observed p-value, and \eqn{P_corrected} is the new p-value after the correction.
The p-value is first computed by 9999 permutations with the function \code{anova.cca} 
(package \code{vegan}), and is then corrected according to the total number of SWMs 
tested (if \code{correction = TRUE}). Although the function can be run without this
correction (\code{correction = FALSE}), using the default value (\code{correction = TRUE})
is strongly recommended to avoid inflated type I error rates (Bauman et al. 2018a).

As a consequence of the necessity of the p-value correction, the significance threshold
decreases as the number of SWMs increases, hence leading to a trade-off between the gain 
of accuracy and the power loss. See \code{Details} of the function 
\code{\link{listw.candidates}} for recommendations about the selection of a 
suitable set of candidate SWMs. 

The optimization criterion of the SWM performed by \code{listw.select} is either based 
on the maximization of the adjusted R2 of all the generated spatial eigenvectors (also 
referred to as spatial predictors or MEM variables) (\code{crit = "global"}), or is based 
on an optimized subset of eigenvectors (\code{crit = "forward"} or \code{"MIR"}).
If the objective is only to optimize the selection of the SWM, without the intervention 
of the selection of a subset of predictors within each SWM (\code{crit = "global"}), then 
the best SWM will be the one maximizing the adjusted global R2, that is, the R2 of the model
of \code{x} against the whole set of generated MEM variables. If \code{autocor = "all"}, 
then the adjusted R2 is computed separately on the MEM associated to positive and negative 
eigenvalues (hereafter positive and negative MEM variables, respectively), and the SWM 
yielding the highest sum of the the two R2 values is selected.
Using \code{crit = "global"} may be interesting when the complete set of MEM variables will
be used, like in Moran spectral randomizations (Wagner and Dray 2015) or when using 
smoothed MEM (Munoz 2009). 

If the MEM variables will be further used in a model including actual predictors 
(e.g. environmental), then a subset of spatial eigenvectors will need to
be selected before proceeding to further analyses to avoid model overfitting and a loss of 
statistical power to detect the contribution of the environment to the 
variability of the response data (Griffith 2003, Dray et al. 2006, Blanchet et al. 2008,
Peres-Neto and Legendre 2010, Diniz-Filho et al. 2012). In this case, the optimization
of the SWM should be based on the best subset of spatial predictors within each
candidate SWM. Although several eigenvector selection approaches have been proposed to 
select this best subset of eigenvectors, Bauman et al. (2018b) showed that two main 
procedures should be preferred, depending on the underlying objective.

The most powerful and accurate selection method, in terms of R2 estimation, is the 
forward selection with double stopping criterion of Blanchet et al. (2008). This method
(\code{crit = "forward"}, default option) should be preferred when the objective is to
capture as accurately as possible the spatial patterns of \code{x}.
If the objective is to optimize the detection of the spatial patterns in the residuals of
a model of the response variable(s) against a set of environmental predictors, for instance,
then \code{x} should be the model residuals, and \code{crit = "forward"}. This can be 
interesting if the objective is to optimize the detection of residual spatial patterns
once the effect of the environmental predictors has been removed.

If however there is no interest in exploring accurate spatial structures in the residuals
of the model of \code{x} against the environmental predictors, and the objective is only
to remove the spatial autocorrelation from the model residuals with a small number of
spatial predictors, then accuracy is not as important and one should focus mainly on the
number of spatial predictors (Bauman et al. 2018b).
In this case, using \code{crit = "MIR"} is the most adapted choice. This optimization 
criterion selects the smallest subset of spatial predictors necessary to capture all the 
spatial autocorrelation in \code{x} (see function \code{\link{MEM.moransel}}. It has the 
advantage to maintain the standard errors of the actual predictor coefficients as low as 
possible. 

If \code{crit = "forward"}, \code{listw.select} performs the forward selection on the 
significant SWMs and selects among these the SWM for which the forward-selected
subset of spatial eigenvectors yields the highest adjusted R2. 
If \code{crit = "MIR"}, \code{listw.select} performs the MIR selection on all the 
significant candidate SWMs, and selects the best SWM as the one with the smallest number 
of MIR-selected spatial eigenvectors. If two or more SWMs present the same smallest number 
of predictors, then the selection is made among them on the basis of the adjusted R2 of the 
selected eigenvectors (as for a same number of spatial eigenvectors, a higher R2 indicates 
a more accurate description of the spatial structure in the residuals (Bauman et al. 2018a)).
If \code{autocor = "all"}, the optimization criteria described above are applied on the
sum of the adjusted R2 or number of selected spatial eigenvectors, for 
\code{crit = "forward"} and \code{"MIR"}, respectively.

Note that the MIR criterion of optimization can only be used for a univariate \code{x}, 
as the Moran's I is a univariate index. If \code{x} is multivariate, then the best 
criterion is the forward selection (see Bauman et al. 2018b).

\code{mem.select} works exactly like \code{listw.select} but for a single SWM. Its purpose
is to test the SWM and, if significant, to optimize the selection of a subset of spatial
predictors within it.
}
\examples{
if(require(spdep)) {
### Create a grid of 15 x 15:
grid <- expand.grid(x = seq(1, 15, 1), y = seq(1, 15, 1))
### Generation of a response variable Y structured at broad scale by linear combination of
### the first three MEM variables to which a normal noise is added:
nb <- cell2nb(nrow = 15, ncol = 15, "queen")
nb2 <- nb2listw(nb, style = "B")
MEM <- scores.listw(nb2, MEM.autocor = "positive")
# Degree of spatial autocorrelation:
intensity <- 0.4
Y_space <- scale(MEM[, 1] + MEM[, 2] + MEM[, 3]) * intensity
Y_noise <- scale(rnorm(n = nrow(MEM), mean = 0, sd = 1)) * (1 - intensity)
Y <- Y_space + Y_noise
### Y is sampled in 100 randomly-chosen sites of the grid:
sample <- sample(c(1:nrow(grid)), 100, replace = FALSE)
xy <- grid[sample, ]
Y_sampled <- Y[sample]
### The function listw.candidates is used to build the spatial weighting matrices that
### we want to test and compare (with the listw.select function). We test a Gabriel's graph, 
### a minimum spanning tree, and a distance-based connectivity defined by a threshold
### distance corresponding to the smallest distance keeping all sites connected (i.e., 
### the defaut value of d2; see help of function listw.candidates). 
### These connectivity matrices are then either not weighted (binary weighting), or 
### weighted by the linearly decreasing function (see help of the function listw.candidates):
candidates <- listw.candidates(coord = xy, gab = TRUE, mst = TRUE, binary = TRUE, 
                               flin = TRUE)
### Number of candidate W matrices generated:
nbw <- length(candidates)
### Significance threshold value after p-value correction (Sidak correction):
1 - (1 - 0.05)^(1/nbw)
### Optimization of the selection of the SWM among the candidates generated above, 
### using the corrected significance threshold calculated above for the global tests:
W_sel <- listw.select(Y_sampled, candidates, autocor = "positive", crit = "forward", 
                    correction = TRUE)
### Some characteristics of the best spatial model:
# Best SWM:
W_sel$best$name
# Selected subset of spatial predictor within the best SWM:
W_sel$best$MEM.select
W_sel$best$NbVar
# Corrected p-value of the global test of the best SWM: 
W_sel$best$pval
# Adjusted R2 of the subset of spatial predictors selected within the chosen SWM:
W_sel$best$R2.select
# p-values of all the tested W matrices:
W_sel$all$pval
# Adjusted R2 of the subset of spatial predictors selected for all the significant
# W matrices:
W_sel$all$R2.select

# See Appendix S3 of Bauman et al. 2018a for more extensive examples and illustrations.
}

}
\references{
Bauman D., Fortin M-J, Drouet T. and Dray S. (2018a) Optimizing the choice of 
a spatial weighting matrix in eigenvector-based methods. Ecology

Bauman D., Drouet T., Dray S. and Vleminckx J. (2018b) Disentangling good from bad 
practices in the selection of spatial or phylogenetic eigenvectors. Ecography, 41, 1--12

Blanchet G., Legendre P. and Borcard D. (2008) Forward selection of explanatory variables.
Ecology, 89(9), 2623--2632

Diniz-Filho J.A.F., Bini L.M., Rangel T.F., Morales-Castilla I. et al. (2012) On the 
selection of phylogenetic eigenvectors for ecological analyses. Ecography, 35, 239--249

Dray S., Legendre P. and Peres-Neto P. R. (2006) Spatial modeling: a comprehensive 
framework for principal coordinate analysis of neighbor matrices (PCNM). Ecological 
Modelling, 196, 483--493

Griffith D. (2003) Spatial autocorrelation and spatial filtering: gaining understanding 
through theory and scientific visualization. Springer, Berlin

Munoz, F. 2009. Distance-based eigenvector maps (DBEM) to analyse metapopulation structure 
with irregular sampling. Ecological Modelling, 220, 2683–-2689

Peres-Neto P. and Legendre P. (2010) Estimating and controlling for spatial structure 
in the study of ecological communities. Global Ecology and Biogeography, 19, 174--184

Sidak Z. (1967) Rectangular confidence regions for the means of 
multivariate normal distributions. Journal of the American Statistical Association, 62(318),
626--633

Wagner H., Dray S. (2015). Generating spatially constrained null models for irregularly 
spaced data using Moran spectral randomization methods. Methods in Ecology and Evolution, 
6, 1169--1178
}
\seealso{
\code{\link{listw.candidates}}, \code{\link{MEM.moransel}}, 
\code{\link{scores.listw}}, \code{\link[vegan]{varpart}}
}
\author{
Bauman David \email{dbauman@ulb.ac.be} or \email{davbauman@gmail.com}
}
\keyword{spatial}
