% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.select.R
\name{mem.select}
\alias{mem.select}
\title{Selection of the best subset of spatial eigenvectors (MEM, Moran's
Eigenvector Maps)}
\usage{
mem.select(x, listw, MEM.autocor = c("positive", "negative", "all"),
  method = c("FWD", "MIR", "global"), MEM.all = FALSE, nperm = 999,
  alpha = 0.05, ...)
}
\arguments{
\item{x}{A vector, matrix, or dataframe of response variable(s). The
\code{method="MIR"} is only implemented for a vector response. Note that
\code{x} can also contain the residuals of a model when other (e.g.,
environmental) variables should be considered in the model.}

\item{listw}{A spatial weighting matrix of class \code{listw}; can be created
with functions of the package \code{spdep}, or with the user-friendly
function \code{listw.candidates}}

\item{MEM.autocor}{Sign of the spatial eigenvectors to generate; "positive",
"negative", or "all", for positively, negatively autocorrelated
eigenvectors, or both, respectively; default is "positive"}

\item{method}{Criterion to select the best subset of MEM variables. Either
\code{forward} (default option), \code{"MIR"} (for univariate \code{x}
only), or \code{"global"} (see \code{Details})}

\item{MEM.all}{A logical indicating if the complete set of MEM variables
should be returned}

\item{nperm}{Number of permutations to perform the tests; Default is 999}

\item{alpha}{Significance threshold value for the tests; Default is 0.05}

\item{\dots}{Other parameters (for internal use with \code{listw.select})}
}
\value{
The function returns a list with: \describe{ \item{global.test}{An
  object of class \code{randtest} containing the result of the global test
  associated to all MEM. If \code{MEM.autocor = "all"}, a list with two
  elements (\code{positive} and \code{negative}) corresponding to the results
  of the global tests performed on positive and negative MEM respectively.}
  \item{MEM.all}{An object of class \code{orthobasisSp} containing the
  complete set of generated MEM variables (generated by
  \code{\link{scores.listw}}. Only returned if \code{MEM.all = TRUE}}
  \item{summary}{A dataframe summarizing the results of the selection
  procedure} \item{MEM.select}{An object of class \code{orthobasisSp}
  containing the subset of significant MEM variables.}}
}
\description{
\code{mem.select} computes the spatial eigenvectors (MEM) of the spatial
weighting matrix (SWM) provided (\code{listw}) and optimize the selection of
a subset of MEM variables relative to response variable(s) stored in
\code{x}. The optimization is done either by maximizing the adjusted
R-squared (R2) of all (\code{method = "global"}) or a subset (\code{method =
"FWD"}) of MEM variables or by minimizing the residual spatial
autocorrelation (\code{method = "MIR"}).
}
\details{
The function provides three different methods to select a subset of
  MEM variables. For all methods, a global test is firstly performed. If
  \code{MEM.autocor = "all"}, two global tests are performed and pvalues are
  corected for multiple comparison (significance threshold equal to
  \code{alpha} divided by two).

  For \code{method = "MIR"}, the global test consists in computing the
  Moran's I of \code{x} (e.g. residuals of the model of the response variable
  against environmental variables) and tests it by permutation (results
  stored in \code{gloabl.test}). If the Moran's I is significant, the
  function performs a selection procedure that searches among the set of
  generated spatial predictors the one that best minimizes the value of the
  Moran's I. A model of \code{x} against the selected eigenvector is built,
  and the significance of the Moran's I of the model residuals is tested
  again. The procedure goes on until the Moran's I of the model residuals is
  not significant anymore, hence the name of Minimization of moran's I in the
  Residuals (MIR).

  For \code{method = "global"} and \code{method = "FWD"}, the global test
  consists in computing the adjusted global R2, that is, the R2 of the model
  of \code{x} against the whole set of generated MEM variables and tests it
  by permutation (results stored in \code{gloabl.test}).

  For \code{method = "global"}, if the adjusted global R2 is significant, the
  functions returns the whole set of generated MEM variables in
  \code{MEM.select}.

  For \code{method = "FWD"}, if the adjusted global R2 is significant, the
  function performs a forward selection procedure that searches among the set
  of generated spatial predictors the one that best maximizes the R2 of the
  model. It is the forward selection with double stopping criterion of
  Blanchet et al. (2008).

  For the \code{method = "FWD"} and \code{method = "MIR"}, the MEM selected
  by the procedure are returned in \code{MEM.select} and a summary of the
  results is provided in \code{summary}. If no MEM are selected, then
  \code{MEM.select} and \code{summary} are not returned.
}
\examples{
if(require(vegan)){ 
# Illustration of the MIR selection on the oribatid mite data
# (Borcard et al. 1992, 1994 for details on the dataset):
# *******************************************************
# Community data (response matrix):
data(mite)
# We will compute the example on a single species:
spe <- mite[, 2]
# Environmental explanatory dataset:
data(mite.env)
# We only use two numerical explanatory variables:
env <- mite.env[, 1:2]
dim(env)
# Coordinates of the 70 sites:
data(mite.xy)
coord <- mite.xy
# We build the model we are interested in:
mod <- lm(spe ~ ., data = env)
# In order to avoid possible type I error rate inflation issues, we check 
# whether the model residuals are independent, and if they are spatially
# autocorrelated, we select a small subset of MEM variables to add to the
# model as covariables with the MIR selection:
# 1) We build a spatial weighting matrix based on Gabriel graph with a
# weighting function decreasing linearly with the distance:
w <- listw.candidates(coord, nb = "gab", weights = "flin")
# 2) We test the spatial autocorrelation of the model residuals and, if
# necessary, select a subset of spatial predictors:
y <- residuals(mod)
MEM <- mem.select(x = y, listw = w[[1]], MEM.autocor = "positive", nperm = 999,
                   alpha = 0.05)
dim(MEM$MEM.select)
# The residuals of the model presented spatial autocorrelation. The selection
# of one MEM variable was sufficient for the Moran's I of the residuals not to
# be significant anymore.
# 3) We can reconstruct our model adding the selected MEM variable as covariables:
env2 <- cbind(env, MEM$MEM.select)
mod_complete <- lm(spe ~ ., data = env2)
summary(mod_complete)$coefficient[, 1]   # Coefficient estimates
summary(mod_complete)$coefficient[, 2]   # Standard errors
}


}
\references{
Bauman D., Drouet T., Dray S. and Vleminckx J. (2018) Disentangling good from
bad practices in the selection of spatial or phylogenetic eigenvectors.
Ecography, 41, 1--12

Blanchet G., Legendre P. and Borcard D. (2008) Forward selection of
explanatory variables. Ecology, 89(9), 2623--2632

Griffith D., Peres-Neto P. (2006) Spatial modeling in Ecology: the
flexibility of eigenfunction spatial analyses. Ecology, 87, 2603--2613
}
\seealso{
\code{\link{listw.candidates}}, \code{\link{listw.select}},
  \code{link{scores.listw}}
}
\author{
David Bauman (\email{dbauman@ulb.ac.be} or
  \email{davbauman@gmail.com}) and Stéphane Dray
}
\keyword{spatial}
