% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.randtest.R
\name{moran.randtest}
\alias{moran.randtest}
\title{Function to compute Moran's index of spatial autocorrelation}
\usage{
moran.randtest(x, listw, nrepet = 999, ...)
}
\arguments{
\item{x}{a \code{vector}, \code{matrix} or \code{data.frame} with numeric data}

\item{listw}{an object of class \code{listw} created for example by \code{\link[spdep]{nb2listw}}}

\item{nrepet}{an integer indicating the number of permutations used in the 
randomization procedure}

\item{\dots}{other arguments to be passed to the \code{\link[ade4]{as.randtest}} or code{\link[ade4]{as.krandtest}} functions.}
}
\value{
An object of class \code{randtest} (one variable) or \code{krandtest} (several variables)
}
\description{
This function computes Moran's I statistic and provide a testing procedure using random permutations.
It is based on the \code{moran.mc} function of the \code{spdep} package. The \code{moran.randtest}
is slightly different as it allows to consider several variables (\code{x} can have more than one columns)
and its ouputs are objects of class \code{randtest} (one variable) or \code{krandtest} (several variables).
}
\examples{

#===========================================
# Computing Moran's I for a set of variables
#===========================================
if(require("ade4", quietly = TRUE)  & require("spdep", quiet = TRUE)){
data(mafragh)
tests <- moran.randtest(mafragh$env, nb2listw(mafragh$nb))
tests
plot(tests)

}

#=============================
# Computing Moran's I for AEMs
#=============================
# Building AEMs
xy <- cbind(1:25,expand.grid(1:5,1:5))
Wdist <- 1/as.matrix(dist(xy[,2:3]))

nb <- cell2nb(5,5,"queen")
bin.mat <- aem.build.binary(nb,xy)
linkBase <- bin.mat[[2]]
link <- linkBase[-which(linkBase[,1] == 0),]
weight <- numeric()

for(i in 1:nrow(link)){
   weight[i] <- Wdist[link[i,1],link[i,2]]
}

AEM <- aem(bin.mat, weight = weight, rm.link0 = TRUE)

# Constructing asymmetric matrix
matasym <- matrix(0,ncol=nsite,nrow=nsite)

for(i in 1:nrow(link)){
    matasym[link[i,1],link[i,2]]<- weight[i]
}

# Build a listw object from the asymmetric matrix
listwAsym <- mat2listw(matasym)

# Calculate Moran's I for AEM
MoranIAEM <- moran.randtest(AEM$vectors, listwAsym)

}
\references{
Moran, P. A. P. (1950). Notes on continuous stochastic phenomena. Biometrika, 37, 17-23.
}
\seealso{
\code{\link[spdep]{moran.mc}}
}
\author{
Stéphane Dray \email{stephane.dray@univ-lyon1.fr}
}
\keyword{spatial}
