% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trichoptera.R
\docType{data}
\name{trichoptera}
\alias{trichoptera}
\title{Trichoptera data set}
\format{A data frame with 220 rows and 58 variables (Site, Date, and 56 species)}
\usage{
data(trichoptera)
}
\description{
A dataset containing the abundances of 56 species in 220 sites
}
\keyword{datasets}
