% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envspace.test.R
\name{envspace.test}
\alias{envspace.test}
\title{Perform a test of the joint space-environment fraction of a variation partitioning
using torus-translation or Moran Spectral Randomisation}
\usage{
envspace.test(E, Ab, coord, MEM_Ab, listw_E, autocor_E = c("positive",
  "negative", "all"), scale = TRUE, regular = FALSE, ntoro = 999,
  nMSR = 999, MSRmethod = "singleton", alpha = 0.05, summary = TRUE)
}
\arguments{
\item{E}{Vector, matrix, or dataframe of environmental variables (rows = sites, 
columns = variables)
           Each variable in E needs to be standardized and normalized prior to the test.}

\item{Ab}{Vector, matrix, or dataframe of species abundances (rows = sites, 
columns = abundances)}

\item{coord}{Matrix or dataframe of spatial coordinates of the sampled sites}

\item{MEM_Ab}{Matrix or dataframe of spatial predictors (MEM variables) for the response
data (\code{Ab})}

\item{listw_E}{An object of class \code{listw} (spatial weights) created by the functions
of the \bold{\code{spdep}} package or returned by the function \code{\link{MEM.modsel}} in
its \code{$listw} output}

\item{autocor_E}{A string indicating whether all the MEM variables built on the basis of
\code{listW_E} should be used to test the significance of a spatial structure in \code{E},
(\code{"all"}), or only those corresponding to positive (\code{"positive"}) or 
negative (\code{"negative"}) autocorrelation; Default is "positive"}

\item{scale}{Logical value indicating whether the environmental variables should be
centered and scaled (standardised data are necessary); Default is \code{TRUE}}

\item{regular}{Logical argument indicating whether a torus-translation test is
performed, additionnaly to the MSR. Set to \code{TRUE} if the sampling design is regular 
(same number of sites on each line, same number of sites on each column), set to
\code{FALSE} otherwise; Default is \code{FALSE}}

\item{ntoro}{Number of permutations performed using torus-translations; Default is 999}

\item{nMSR}{Number of permutations performed using MSR; Default is 999}

\item{MSRmethod}{Algorithm of \code{\link{msr}} to be used to perform the MSR. The
three available procedures are "singleton" (default), "pair", and "triplet" (see
\code{\link{msr}} for details)}

\item{alpha}{Threshold value of null hypothesis rejection for the test of a 
spatial structure in the environment, and for the joint environment-space fraction of 
the VP; Default is 0.05}

\item{summary}{Logical; Whether a message summarising the results should be returned
or not; Default is \code{TRUE}}
}
\value{
The function returns a list containing the following elements: \describe{
\item{R2adj}{The adjusted R-squared value of the JSEF.}
\item{pval_TT}{The significance value of the JSEF obtained by TT.}
\item{pval_MSR}{The significance value of the JSEF obtained by MSR.}
}
}
\description{
This function performs a test of the joint space-environment fraction (JSEF) of a 
variation partitioning (VP). The adjusted R-squared (Peres-Neto et al. 2006; R2adj) of 
the JSEF is not an actual R2, as it is computed by subtracting the R2adj of other
fractions of the VP and therefore has zero degree of freedom. The JSEF
can therefore not be computed in the classical way. The function \code{envspace.test}
provides two ways of testing this fraction: a torus-translation (or torus randomisation)
test (TT), for regular sampling designs, and Moran spectral randomization (MSR). The 
function first checks whether the both the environment and the response data display 
significant spatial structures, and then proceeds to the significance test if and only 
if both datasets display a spatial structure.
}
\examples{
if(require(vegan)) { 
# Illustration of the test of the JSEF on the oribatid mite data
# (Borcard et al. 1992, 1994 for details on the dataset):
# Community data (response matrix):
data(mite)
# Hellinger-transformation of the community data (Legendre and Gallagher 2001):
Y <- decostand(mite, method = "hellinger")
# Environmental explanatory dataset:
data(mite.env)
# For the example, we only use two numerical explanatory variables:
env <- mite.env[, 1:2]
dim(Y)
dim(env)
# Coordinates of the 70 sites:
data(mite.xy)
coord <- mite.xy

### Building a list of spatial weighting matrix (W matrix) candidates for the 
### optimisation of the W matrix selection for 'Y' and 'env':
# We create five candidates: a connectivity matrix based on a Gabriel graphs, on
# a minimum spanning tree (i.e., two contrasted graph-based W matrices), either
# not weighted, or weighted by a linear function decreasing with the distance),
# and a distance-based W matrix corresponding to the connectivity and weighting
# criteria of the original PCNM method:
candidates <- listw.candidates(coord, del = FALSE, gab = TRUE, rel = FALSE,
                               mst = TRUE, DB = FALSE, PCNM = TRUE, 
                               binary = TRUE, flin = TRUE, fconcdown = FALSE,
                               fconcup = FALSE)
### Optimisation of the choice of a W matrix:
# W matrix for 'Y':
modsel_Y <- MEM.modsel(Y, candidates, correction = TRUE, autocor = "positive")

paste("The best W matrix for 'Y' was ", modsel_Y$name, ". The forward selection ",
      "with double stopping criterion (Blanchet et al. 2008) selected a best subset",
      " of spatial predictors within this W matrix. This subset contains ", 
      modsel_Y$NbVar, " MEM variables and has an adjusted R2 of ", 
      round(modsel_Y$R2adj, 3), ".", sep = "")

modsel_env <- MEM.modsel(env, candidates, correction = TRUE, autocor = "all")

paste("The best W matrix for 'env' was ", modsel_env$name, ". The forward selection ",
      "with double stopping criterion (Blanchet et al. 2008) selected a best subset",
      " of spatial predictors within this W matrix. This subset contains ", 
      modsel_env$NbVar, " MEM variables and has an adjusted R2 of ", 
      round(modsel_env$R2adj, 3), ".", sep = "")

### We perform the variation partitioning:
# Subset of selected MEM variables within the best W matrix:
MEM_Ab <- modsel_Y$MEM.select

VP <- varpart(Y, env, MEM_Ab)
plot(VP)

# Test of the joint space-environment fraction (fraction [b]):
JSEF.test <- envspace.test(E = env, Ab = Y, coord = coord, MEM_Ab = MEM_Ab,
                           listw_E = modsel_env$listw, scale = TRUE, regular = FALSE,
                           nMSR = 999)
JSEF.test$R2adj
JSEF.test$pval_MSR

# The JSEF is very highly significant (p-value = 0).
}

}
\references{
Bauman D., Drouet T., Dray S. and Vleminckx J. (2017) Disentangling good from bad 
practices in the selection of spatial or phylogenetic eigenvectors. Ecography 

Blanchet G., Legendre P. and Borcard D. (2008) Forward selection of explanatory variables.
Ecology, 89(9), 2623--2632

Legendre P., Gallagher E.D. (2001) Ecologically meaningful transformations for 
ordination of species data. Oecologia, 129(2), 271--280

Peres-Neto P., Legendre P., Dray S., Borcard D. (2006) Variation partitioning of 
species data matrices: estimation and comparison of fractions. Ecology, 87(10), 
2614--2625

Peres-Neto P. and Legendre P. (2010) Estimating and controlling for spatial structure 
in the study of ecological communities. Global Ecology and Biogeography, 19, 174--184
}
\seealso{
\code{\link{varpart}}, \code{\link{MEM.modsel}}, \code{\link{listw.candidates}}
}
\author{
Jason Vleminckx and David Bauman, \email{jasvlx86@gmail.com}, 
\email{davbauman@gmail.com}
}
\keyword{spatial}
