% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listw.candidates.R
\name{listw.candidates}
\alias{listw.candidates}
\title{Function to create a list of spatial weighting matrices}
\usage{
listw.candidates(coord, style = "B", del = TRUE, gab = TRUE, rel = TRUE,
  mst = TRUE, PCNM = TRUE, DB = TRUE, DBthresh, binary = TRUE,
  flin = TRUE, fconcdown = TRUE, fconcup = TRUE, y_fconcdown = 5,
  y_fconcup = 0.5)
}
\arguments{
\item{coord}{Vector, matrix, or dataframe of point coordinates}

\item{style}{Coding scheme style (see \code{nb2listw} of the \code{spdep}
package). Can take values 'W', 'B', 'C', 'U', 'minmax', and 'S'; default is
'B'}

\item{del}{Defines whether a B matrix based on the Delaunay triangulation
should be used; default is TRUE. No edge effect correction implemented here}

\item{gab}{Defines whether a B matrix based on a Gabriel's graph should be
used; default is TRUE}

\item{rel}{Defines whether a B matrix based on a relative neighbourhood graph
should be used; default is TRUE}

\item{mst}{Defines whether a B matrix based on a minimum spanning tree should
be used; default is TRUE}

\item{PCNM}{Defines whether a distance-based W matrix based on the principal 
coordinates of neighbour matrices (PCNM) criteria should be used (see
'Details'); default is TRUE}

\item{DB}{Defines whether a distance-based W matrix should be built; default
is TRUE}

\item{DBthresh}{Only considered if DB is TRUE; defines the connectivity
distance threshold below which two sites are connected. It can be a single
value or a vector of values, in which case a different W matrix will be
generated for each threshold value. The default value is the minimum
distance keeping all points connected (i.e., the largest edge of the 
minimum spanning tree)}

\item{binary}{Defines whether W matrices based on the selected B matrices
should be created without weights on the connexions; default is TRUE}

\item{flin}{Defines whether the linear weighting function should be used (see
Details below); default is TRUE}

\item{fconcdown}{Defines whether the concave-down weighting function should
be used (see Details below); default is TRUE}

\item{fconcup}{Defines whether the concave-up weighting function should be
used (see Details below); default is TRUE}

\item{y_fconcdown}{Single value or vector of values of the \code{y} parametre
in the concave-down weighting function; default is 5}

\item{y_fconcup}{Single value or vector of values of the \code{y} parametre
in the concave-up weighting function; default is 0.5}
}
\value{
A list of W matrices. Each element of the list was built by
  \code{nb2listw} (package \code{spdep}) and therefore is of class
  \code{listw} and \code{nb}. The name of each element of the list (W matrix)
  is composed of the corresponding B and A matrices, followed (if any) by the
  \code{y} parametre value of the weighting function.
}
\description{
This function is a user-friendly way to create a list of spatial weighting
matrices (W matrices) by selecting a set of predefined connectivity and
weighting matrices (B and A matrices, respectively). The list can then be fed
to the function \code{MEM.modsel} to optimise the selection of the W matrix
and select the best eigenvector subset within this matrix while controlling
the type I error rate.
}
\details{
The function allows to construct W matrices based on any combination
  of B and A matrices. The B matrices are either graph-based or
  distance-based. The function proposes the Delaunay triangulation, Gabriel
  graph, relative neighbourhood graph, and the minimum spanning tree criteria
  to build a graph-based B matrix. Distance-based W matrices can be built
  with the principal coordinates of neighbour matrices (PCNM; Borcard and
  Legendre 2002) criteria (see details below), or using another threshold 
  distance to define the connected site pairs. The A matrix can be based on a
  binary, linear, concave-down, or concave-up function. The linear,
  concave-down, and concave-up weighting functions are defined by \eqn{1 -
  (D/dmax)}, \eqn{1 - (D/dmax)^y}, and \eqn{1 / D^y}, respectively, where
  \code{D} is the euclidean distance between the two sites considered,
  \code{dmax} is the maximum euclidean distance between two sites, and 
  \code{y} is a user-defined parametre that can either be a single value or a
  vector of values. The argument \code{PCNM} consists in constructing a
  distance-based W matrix based on the largest edge of the minimum spanning
  tree as a connectivity distance threshold, and then by weighting the links
  by the function \eqn{1-(D/(4*t))^2}, where \code{D} is the euclidean
  distance between the sites, and \code{t} is the distance threshold below
  which two sites are considered connected (Dray et al. 2006). As optimising
  the choice of a W matrix has to be done with a p-value correction depending
  on the number of W matrix candidates tested (see function
  \code{MEM.modsel}), Bauman et al. (2018) strongly encouraged plotting the
  concave-down and concave-up weighting functions with several parametre
  values in order to only choose the realistic ones to build the W matrix
  candidates (e.g., ranging between 0.1 and 1 for the concave-up function, as
  values over 1 would make no ecological sense). First visualising the 
  connectivity schemes with the \code{createlistw} function may also help
  choosing the B matrices to select for the \code{listw.candidates} function.
}
\examples{
### Create 100 random sampling locations in a squared grid of 120 x 120:
xy <- matrix(nrow = 100, ncol = 2)
xy[, 1] <- sample(c(1:120), 100, replace = FALSE)
xy[, 2] <- sample(c(1:120), 100, replace = FALSE)
### The function listw.candidates is used to build the spatial weighting matrices that
### we want to test and compare (with the MEM.modsel function). We test a Gabriel's graph, 
### a minimum spanning tree, and a distance-based connectivity defined by a threshold
### distance corresponding to the smallest distance keeping all sites connected (i.e., 
### the defaut value of DBthresh). These connectivity matrices are then either not weighted 
### (binary weighting), or weighted by the linearly decreasing function:
candidates <- listw.candidates(coord = xy, del = FALSE, rel = FALSE, fconcdown = FALSE, 
                               fconcup = FALSE, PCNM = FALSE)
### Construction of a different list of spatial weighting matrices. This time, the
### connexions are defined by a distance-based criterion based on the same threshold
### value, but the connexions are weighted by the concave-down function with a y parametre
### varying between 2 and 5, and a concave-up function with a y parametre of 0.2.
candidates2 <- listw.candidates(coord = xy, del = FALSE, gab = FALSE, rel = FALSE, 
                                mst = FALSE, PCNM = FALSE, bin = FALSE, y_fconcdown = c(1:5),
                                y_fconcup = 0.2)
### Number of spatial weighting matrices generated:
length(candidates2) 

}
\references{
Bauman D., Fortin M-J, Drouet T. and Dray S. (2018a) To link or not to link: 
optimising the choice of a spatial weighting matrix in eigenvector-based methods. Ecology

Borcard D. and Legendre P. (2002) All-scale spatial analysis of 
ecological data by means of principal coordinates of neighbour matrices.
Ecological Modelling, 153, 51--68
  
Dray S., Legendre P. and Peres-Neto P. R. (2006) Spatial modeling: a
comprehensive framework for principal coordinate analysis of neighbor
matrices (PCNM). Ecological Modelling, 196, 483--493
}
\seealso{
\code{\link{createlistw}}, \code{\link{MEM.modsel}}
}
\author{
Bauman David \email{dbauman@ulb.ac.be} or \email{davbauman@gmail.com}
}
\keyword{spatial}
