% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM.modsel.R
\name{MEM.modsel}
\alias{MEM.modsel}
\title{Function to optimize the selection of a spatial weighting matrix and select the best 
subset of eigenvectors within it}
\usage{
MEM.modsel(x, candidates, autocor = c("positive", "negative", "all"),
  crit = c("selection", "global"), alpha_thresh = 0.05, correction = TRUE,
  summary = TRUE)
}
\arguments{
\item{x}{Vector, matrix, or dataframe of response data}

\item{candidates}{A list of one or more spatial weighting matrices of the class 
\code{listw}; \code{candidates} can be a list containing one or several spatial 
weighting matrices, but it can also be a single listw object.}

\item{autocor}{Sign of the spatial eigenvectors to consider; "positive", "negative", 
or "all", for positively, negatively autocorrelated eigenvectors, or both, respectively; 
default is "positive"}

\item{crit}{Criterion to select the best W matrix among the set of significant candidate 
matrices. The default option (\code{selection}) compares the adjusted R2 of the forward
selected subsets of spatial eigenvectors and defines the optimized W matrix as the one
corresponding to the highest R2 value. If one wants to select the best W matrix on the basis
of all the MEM variables generated by \code{scores.listw} (see argument \code{autocor}), then
\code{crit} can be changed to \code{global}.}

\item{alpha_thresh}{Uncorrected predefined significance value; default is 0.05}

\item{correction}{wheter the p-value of the global test performed on each W matrix should
be corrected for multiple tests (TRUE) or not (FALSE); default is TRUE}

\item{summary}{Logical; Whether a message summarizing the results should be returned
or not; The R2 value returned in the message depends on
the argument \code{crit} (R2 of the subset of forward selected predictors or R2 of the
global model); Default is \code{TRUE}}
}
\value{
The function returns two lists containing several features of the W matrix selected
by the optimization procedure, and general features of all the W matrices generated and 
tested: 
The first list, \code{all}, contains: \describe{
\item{MEM.all}{A list of all the MEM variables (as generated by \code{scores.listw}) 
for the W matrices (\code{listw} objects) provided to the function, in the same order
as the W matrices in \code{candidates}.}
\item{pval}{The p-values (corrected by the Sidak correction if \code{correction = TRUE})
of the global models for each W matrix tested.}
\item{R2.global}{The adjusted R2 of the complete set of MEM variables of each W 
matrix. R2.global are available only for the significant W matrices.}
}

The second list, \code{best}, contains the features of the best model: \describe{
\item{MEM.all}{Dataframe of the complete set of eigenvectors of the best W matrix.}
\item{MEM.select}{Dataframe of the eigenvectors selected by the forward selection. These
are the spatial predictors that should be used for further analyses. In the paradigm of
'spatial legacy' (Peres-Neto and Legendre 2010), these spatial predictors are used to
model the spatial patterns of the response data as accurately as possible. For instance,
using them together with a set of environmental variables in a variation partitioning 
(see function \code{\link{varpart}} from the \code{vegan} package) allows assessing the
pure and joint contribution of the environmental and spatial predictors. In the 
'spatial nuisance' paradigm (Peres-Neto and Legendre 2010), the spatial predictors are
used to remove the spatial signal from the response data in order to respect the condition
of independency of the residuals in classical OLS or GLM models (spatial filtering through
spatial eigenvector mapping for instance, Diniz-Filho et al. 2005). See Bauman et al.
(2018b) for a discussion on the choice of the spatial eigenvector selection method 
according to the univariate or multivariate nature of the response data, and to the
spatial paradigm considered.} 
\item{listw}{Element of \code{candidates} (an object of class \code{listw}) corresponding 
to the best W matrix.}
\item{MEM.AdjR2Cum}{Vector of the cumulative adjusted R2 of the eigenvectors of 
\code{MEM.select}.}
\item{name}{Name of the best W matrix ("Connectivity matrix_Weighting matrix") if the 
list of W matrix candidates was generated by the function \code{\link{listw.candidates}}. 
Otherwise, corresponds to the index of the best W matrix in the list \code{candidates} 
provided to \code{MEM.modsel}.} 
\item{pval}{Global p-value of the best W matrix (after multiple-test correction, if 
the \code{correction} argument = TRUE).}
\item{R2.global}{The adjusted R2 of the complete set of MEM variables of the best W 
matrix.}
\item{R2.select}{Adjusted R2 of \code{MEM.select} (the subset of forward selected MEM
variables).}
\item{NbVar}{Number of spatial predictors selected by the forward selection.}
\item{bestw_index}{Index of the best W matrix in the list \code{candidates} provided 
to \code{MEM.modsel}.}
}
If \code{autocor = "all"} and a significant spatial model was selected for both the
MEM variables associated to positive and negative eigenvalues, then both lists
(\code{all} and \code{best}) are generated twice, for both sets of MEM variables,
and are returned in two different lists: \describe{
\item{MEM.pos}{Lists \code{all} and \code{best} for the MEM variables corresponding to
a positive autocorrelation.}
\item{MEM.neg}{Lists \code{all} and \code{best} for the MEM variables corresponding to
a negative autocorrelation.}
}
}
\description{
\code{MEM.modsel} computes spatial eigenvectors for various definitions of spatial weighting
matrices (W matrices) and optimizes the selection of the W matrix by maximizing the
value of the adjusted R-squared (R2) while controling the type I error rate (see
Bauman et al. 2018a). The function also selects the best subset of eigenvectors to be
used as spatial predictors within the best W matrix by performing a forward selection with
double stopping criterion (Blanchet et al. 2008). The best W matrix returned is selected
on the basis of the R2 of either the global models or the subsets of eigenvectors selected
by the forward selections.
\code{MEM.modsel} combines calls to the functions \code{scores.listw} and \code{forward.sel}.
The list of W matrix candidates can easily be generated using the user-friendly 
\code{listw.candidates} function. The significance of each W matrix is tested by
9999 permutations with the function \code{anova.cca} (package \code{vegan}).
}
\details{
While the selection of the W matrix is the most critical step of the spatial
eigenvector-based methods (Dray et al. 2006), Bauman et al. (2018a) showed that 
optimizing the choice of the W matrix led to inflated type I error rates if an
explicit control of the number of W matrices tested was not applied. The function
\code{MEM.modsel} therefore applies a Sidak correction (Sidak 1967) for multiple tests to
the p-value of the global test of each W matrix (i.e., the model integrating the 
whole set of spatial predictors). The Sidak correction is computed as: 
\eqn{P_corrected = 1 - (1 - P)^n}, where \eqn{n} is the number of tests performed, \eqn{P}
is the observed p-value, and \eqn{P_corrected} is the new p-value after the correction.
The p-value is first computed by 9999 permutations with the function \code{anova.cca} 
(package \code{vegan}), and is then corrected according to the total number of W matrices 
tested (if \code{correction = TRUE}). Although the function can be run without this
correction (\code{correction = FALSE}), using the default value (\code{correction = TRUE})
is strongly recommended to avoid highly inflated type I error rates (Bauman et al. 2018).
As a consequence of the necessity of the p-value correction, the significance threshold
decreases as the number of W matrices tested increases, hence leading to a trade-off
between the gain of accuracy and the statistical power loss. See 'Details' of the 
function \code{\link{listw.candidates}} for recommendations about selecting a suitable set 
of candidate W matrices. 
Once the W matrix was optimized, a subset of spatial eigenvectors (also referred to as
spatial predictors or MEM variables) should be selected before proceding to further 
analyses in order both to avoid model overfitting and a loss of statistical power to detect 
the contribution of the environment to the variability of the response data 
(Griffith 2003, Dray et al. 2006, Blanchet et al. 2008, Peres-Neto and Legendre 2010, 
Diniz-Filho et al. 2012). Although several eigenvector selection approaches have been 
proposed, Bauman et al. (2018b) showed that the most powerful and accurate method, 
in terms of adjusted R2 estimation, was the forward selection with double stopping 
criterion of Blanchet et al. (2008). 
The function \code{MEM.modsel} performs the forward selection on the significant W 
matrices and selects among these the best W matrix as the one for which the 
eigenvector subset resulting from the forward selection displays the highest adjusted R2.
\code{MEM.modsel} can also return the W matrix yielding the highest global R2 (i.e., 
considering the models of the response variable against the entire set of MEM variables
generated for a given W matrix) as the best W matrix (see argument \code{crit}).
}
\examples{
if(require(spdep)) {
### Create a grid of 15 x 15:
grid <- expand.grid(x = seq(1, 15, 1), y = seq(1, 15, 1))
### Generation of a response variable Y structured at broad scale by linear combination of
### the first three MEM variables to which a normal noise is added:
nb <- cell2nb(nrow = 15, ncol = 15, "queen")
nb2 <- nb2listw(nb, style = "B")
MEM <- scores.listw(nb2, MEM.autocor = "positive")
# Degree of spatial autocorrelation:
intensity <- 0.4
Y_space <- scale(MEM[, 1] + MEM[, 2] + MEM[, 3]) * intensity
Y_noise <- scale(rnorm(n = nrow(MEM), mean = 0, sd = 1)) * (1 - intensity)
Y <- Y_space + Y_noise
### Y is sampled in 100 randomly-chosen sites of the grid:
sample <- sample(c(1:nrow(grid)), 100, replace = FALSE)
xy <- grid[sample, ]
Y_sampled <- Y[sample]
### The function listw.candidates is used to build the spatial weighting matrices that
### we want to test and compare (with the MEM.modsel function). We test a Gabriel's graph, 
### a minimum spanning tree, and a distance-based connectivity defined by a threshold
### distance corresponding to the smallest distance keeping all sites connected (i.e., 
### the defaut value of d2; see help of function listw.candidates). 
### These connectivity matrices are then either not weighted (binary weighting), or 
### weighted by the linearly decreasing function (see help of the function listw.candidates):
candidates <- listw.candidates(coord = xy, gab = TRUE, mst = TRUE, binary = TRUE, 
                               flin = TRUE)
### Number of W matrix candidates generated:
nbw <- length(candidates)
### Significance threshold value after p-value correction (Sidak correction):
1 - (1 - 0.05)^(1/nbw)
### Optimization of the selection of the W matrix among the candidates generated above, 
### using the corrected significance threshold calculated above for the global tests:
W_sel <- MEM.modsel(Y_sampled, candidates, autocor = "positive", crit = "selection", 
                    correction = TRUE)
### Some characteristics of the best spatial model:
# Best W matrix:
W_sel$best$name
# Selected subset of spatial predictor within the best W matrix:
W_sel$best$MEM.select
W_sel$best$NbVar
# Corrected p-value of the global test of the best W matrix: 
W_sel$best$pval
# Adjusted R2 of the subset of spatial predictors selected within the chosen W matrix:
W_sel$best$R2.select
# p-values of all the tested W matrices:
W_sel$all$pval
# Adjusted R2 of the subset of spatial predictors selected for all the significant
# W matrices:
W_sel$all$R2.select
}
 
}
\references{
Bauman D., Fortin M-J, Drouet T. and Dray S. (2018a) Optimizing the choice of 
a spatial weighting matrix in eigenvector-based methods. Ecology

Bauman D., Drouet T., Dray S. and Vleminckx J. (2018b) Disentangling good from bad 
practices in the selection of spatial or phylogenetic eigenvectors. Ecography, 41, 1--12

Blanchet G., Legendre P. and Borcard D. (2008) Forward selection of explanatory variables.
Ecology, 89(9), 2623--2632

Diniz-Filho J.A.F., Bini L.M., Rangel T.F., Morales-Castilla I. et al. (2012) On the 
selection of phylogenetic eigenvectors for ecological analyses. Ecography, 35, 239--249

Dray S., Legendre P. and Peres-Neto P. R. (2006) Spatial modeling: a comprehensive 
framework for principal coordinate analysis of neighbor matrices (PCNM). Ecological 
Modelling, 196, 483--493

Griffith D. (2003) Spatial autocorrelation and spatial filtering: gaining understanding 
through theory and scientific visualization. Springer, Berlin

Peres-Neto P. and Legendre P. (2010) Estimating and controlling for spatial structure 
in the study of ecological communities. Global Ecology and Biogeography, 19, 174--184

Sidak Z. (1967) Rectangular confidence regions for the means of 
multivariate normal distributions. Journal of the American Statistical Association, 62(318),
626--633
}
\seealso{
\code{\link{listw.candidates}}, \code{\link{scores.listw}}, 
\code{\link[vegan]{varpart}}
}
\author{
Bauman David \email{dbauman@ulb.ac.be} or \email{davbauman@gmail.com}
}
\keyword{spatial}
